move all from strings in device.py to their own python submodule
