+++ belay/snippets/convenience_imports.py
@@ -1,0 +1,4 @@
+import binascii, errno, hashlib, machine, os, time
+from machine import ADC, I2C, Pin, PWM, SPI, Timer
+from time import sleep
+from micropython import const
