+++ belay/snippets/try_mkdir.py
@@ -1,0 +1,6 @@
+import os, errno
+try:
+    os.mkdir('%s')
+except OSError as e:
+    if e.errno != errno.EEXIST:
+        raise
